% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWindowPictureFuncs.R
\name{extendGffTrans}
\alias{extendGffTrans}
\title{An example how to adopt the gff file to be used with cleavageWindows}
\usage{
extendGffTrans(gffTrans = gffTrans)
}
\arguments{
\item{gffTrans}{Path to the gff file}
}
\description{
An example how to adopt the gff file to be used with cleavageWindows
}
\examples{
extendGffTrans(gffTrans)
}
\keyword{extendGffTrans}
