% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleavageModelAndPrediction.R
\name{kerasCreateModelCNN_2d}
\alias{kerasCreateModelCNN_2d}
\title{Function used to create a 2d CNN model}
\usage{
kerasCreateModelCNN_2d(
  denseLoop = 3,
  testPics,
  trainLabels,
  epochs,
  batch_size,
  dropout = 0.2,
  patience = 0,
  validation_split,
  activation = "relu",
  activationFinal = "softmax",
  pixels = 28,
  optimizer = "adam",
  convolutionalLoop = 2,
  NO_pooling = 1
)
}
\arguments{
\item{denseLoop}{Number of times to iterate the dense layers loops}

\item{testPics}{Image info object generated by kerasCreateDataset_2d}

\item{trainLabels}{trainLabels object generated by}

\item{epochs}{Number of epochs in keras}

\item{batch_size}{Batch size in keras}

\item{dropout}{Keras dropout}

\item{patience}{EarlyStopping patience}

\item{validation_split}{Fraction of the dataset that will be used for validation}

\item{activation}{Standard function activation}

\item{activationFinal}{Final activation function of the model}

\item{pixels}{Pixel size of the images, assigned in kerasCreateDataset_2d}

\item{optimizer}{Optimizer}

\item{convolutionalLoop}{Number of times to iterate the convolutional layers loop}

\item{NO_pooling}{Number of times to perform pooling in the convolutional layers loop (max = 2)}
}
\description{
The model is used to train a Convolutional neural network model that can recognize images
}
\examples{
kerasCreateModelCNN_2d(testPics,
trainLabels,
denseLoop = 3,
epochs,
batch_size,
dropout = 0.2,
patience = 0,
validation_split,
activation = "relu",
activationFinal = "softmax",
pixels =28,
optimizer = "adam",
convolutionalLoop = 2,
NO_pooling = 1
)
}
\keyword{kerasCreateModelCNN}
