% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleavageModelAndPrediction.R
\name{kerasListTrue}
\alias{kerasListTrue}
\title{Constructs a list of the true cleavages}
\usage{
kerasListTrue(pathToTrue = "data/example/", saveDirLevelNr = 3)
}
\arguments{
\item{pathToTrue}{Path to where to look for true cleavages (dir called true) recursively}

\item{saveDirLevelNr}{Depth of recursion where the true dirs should be found at}
}
\description{
Searches for the true cleavages in dirs called true at the depth of recursion
The list can be used to filer away false cleavages from the original dataset
}
\examples{
kerasListTrue(pathToTrue = "keras/3categories/",
saveDirLevelNr = 3 )
}
\keyword{kerasListTrue}
