% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWindowPictureFuncs.R
\name{makeGFF}
\alias{makeGFF}
\title{Constructs a gff object}
\usage{
makeGFF(genome, transcript, outDir)
}
\arguments{
\item{genome}{Path to genome}

\item{transcript}{Path to transcripts}

\item{outDir}{Path to dir where the gff will be written}
}
\description{
Reqires spaln installed on your system, a genome and transcripts
}
\examples{
makeGFF(genome,transcript, outDir)
}
\keyword{makeGFF}
