% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWindowPictureFuncs.R
\name{readGFF}
\alias{readGFF}
\title{Read a gff object int R}
\usage{
readGFF(path)
}
\arguments{
\item{path}{Path to the gff file}
}
\description{
Read a gff object int R
}
\examples{
readGFF(path)
}
\keyword{readGFF}
