% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleavageModelAndPrediction.R
\name{runCLR}
\alias{runCLR}
\title{Optimizes the cyclical learning rate (CLR) based convolutional neural network (CNN) model}
\usage{
runCLR(
  optimizer2 = keras::optimizer_sgd(lr = lr_max, decay = 0),
  unitPower2,
  trainLabels2 = trainLabels,
  epochs2 = 5,
  batch_size2,
  dropout2 = 0.2,
  patience2 = 30,
  validation_split2 = 0.2,
  NOfilters2,
  NO_pooling2 = 1,
  testPics2 = testPics,
  Learning_rate_l2 = Learning_rate_l,
  Learning_rate_h2 = Learning_rate_h,
  pathOut = "data/example/bayesmodels/"
)
}
\arguments{
\item{optimizer2}{Optimizer}

\item{unitPower2}{Number of times to iterate the dense layers loops}

\item{trainLabels2}{trainLabels object generated by}

\item{epochs2}{Number of epochs in keras}

\item{batch_size2}{Batch size in keras}

\item{dropout2}{Keras dropout}

\item{patience2}{EarlyStopping patience}

\item{validation_split2}{Fraction of the dataset that will be used for validation}

\item{NOfilters2}{Number of times to iterate the convolutional layers loop}

\item{NO_pooling2}{Number of times to perform pooling in the convolutional layers loop (max = 2)}

\item{testPics2}{Image info object generated by kerasCreateDataset_2d}

\item{Learning_rate_l2}{Minimum CLR assigned from tuneCLR}

\item{Learning_rate_h2}{Maximum CLR assigned from tuneCLR}

\item{pathOut}{path where the generated model(s) will be saved}
}
\description{
The function optimizes the cyclical learning rate (CLR) based convolutional neural network (CNN) model based on the hyperparameters assigned
}
\examples{
runCLR(optimizer2 = optimizer_sgd(lr=lr_max, decay=0),
unitPower2,
trainLabels2 = trainLabels,
epochs2 = 5,
batch_size2,
dropout2 = 0.2,
patience2 = 30,
validation_split2 = 0.2,
NOfilters2 ,
NO_pooling2 = 1,
testPics2 = testPics,
Learning_rate_l2 = Learning_rate_l,
Learning_rate_h2 = Learning_rate_h,
pathOut = "keras/3categories/bayesmodels/"
)
}
\keyword{runCLR}
