% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleavageModelAndPrediction.R
\name{smartPARE_ListTrue}
\alias{smartPARE_ListTrue}
\title{Constructs a list of the true cleavages}
\usage{
smartPARE_ListTrue(pathToTrue = "data/example/", recursionLevel = 0)
}
\arguments{
\item{pathToTrue}{Path to where to recursively look for true cleavages (dirs called true)}

\item{recursionLevel}{Depth of recursion where the true dirs should be found at}
}
\description{
Searches for the true cleavages in dirs called true at the depth of recursion
The list can be used to filer away false cleavages from the original dataset
}
\examples{
smartPARE_ListTrue(pathToTrue = "keras/3categories/",
recursionLevel = 3 )
}
\keyword{smartPARE_ListTrue}
