% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleavageModelAndPrediction.R
\name{smartPARE_train}
\alias{smartPARE_train}
\title{Train the smartPARE CNN model}
\usage{
smartPARE_train(
  homePath1 = "example/",
  pixels1 = 28,
  search_bound = list(denseLoop2 = c(0, 4), epochs2 = c(100, 300), batch_size2 = c(32,
    128), dropout2 = c(0, 0.3), validation_split2 = c(0.1, 0.4), convolutionalLoop2 =
    c(1, 4), NO_pooling2 = c(1, 2)),
  n_iter = 100
)
}
\arguments{
\item{homePath1}{Path to directory containing a directory with the following subdirs train/goodUp (true cleavages on 5' strand), train/goodDown (true cleavages on 3' strand) and train/bad (false cleavages)}

\item{pixels1}{Number of pixels to convert each image to}

\item{search_bound}{List of min and max values for the following variables: denseLoop2, epochs2, batch_size2, validation_split2, convolutionalLoop2 and NO_pooling2}

\item{n_iter}{Number of iterations to run the Bayesian optimization}
}
\description{
Build the CNN model based on directories with cleavage images; train/goodUp (true cleavages on 5' strand), train/goodDown (true cleavages on 3' strand) and train/bad (false cleavages)
}
\examples{
smartPARE_train(homePath1 = "example/",
pixels1 = 28,
search_bound = list(denseLoop2 = c(0,4),  
                    epochs2 = c(100, 300),  
                    batch_size2 = c(32,128),  
                  dropout2 = c(0, 0.3),  
                    validation_split2 = c(0.1,0.4),  
                 convolutionalLoop2 = c(1,4),  
                 NO_pooling2 = c(1,2)  
),
n_iter = 100)
}
\keyword{smartPARE_train}
