% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleavageModelAndPrediction.R
\name{smartPAREparse}
\alias{smartPAREparse}
\title{Parses the smartPARE output file into a dataset}
\usage{
smartPAREparse(
  smartPAREresultFile = paste0(homePath1, "example_results/true.txt")
)
}
\arguments{
\item{smartPAREresultFile}{smartPARE output filepath}
}
\description{
Parses the smartPARE output file into a dataset based on the path of each true image
}
\examples{
smartPAREparse(smartPAREresultFile = paste0(homePath1,"example_results/true.txt"))

}
\keyword{smartPAREparse}
